requirements.txt
openai-whisper>=20230918
jiwer>=3.0.3
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
torch>=2.0.0
